vinylcontrol/vinylcontrolxwax: Use float math

+++ src/vinylcontrol/vinylcontrolxwax.cpp
@@ -839,7 +839,7 @@ float VinylControlXwax::getAngle() {
         return -1.0;
     }
 
-    double rps = timecoder_revs_per_sec(&timecoder);
+    float rps = timecoder_revs_per_sec(&timecoder);
     // Invert angle to make vinyl spin direction correct.
-    return 360 - (static_cast<int>(pos / 1000.0 * 360.0 * rps) % 360);
+    return 360 - (static_cast<int>(pos / 1000.0f * 360.0f * rps) % 360);
 }
